feat : `ErrorCode` 인터페이스로 리팩토링
