+++ src/main/java/nextstep/subway/line/dto/path/PathRequest.java
@@ -1,7 +1,7 @@
 package nextstep.subway.line.dto.path;
 
 import java.util.Objects;
-import nextstep.subway.common.exception.ErrorCode;
+import nextstep.subway.common.exception.CommonErrorCode;
 import nextstep.subway.common.exception.InvalidParameterException;
 
 public class PathRequest {
@@ -17,11 +17,11 @@ public class PathRequest {
 
     private void validEmpty(Long source, Long target) {
         if (Objects.isNull(source) || Objects.isNull(target)) {
-            throw InvalidParameterException.of(ErrorCode.NOT_EMPTY);
+            throw InvalidParameterException.of(CommonErrorCode.NOT_EMPTY);
         }
 
         if (source.equals(target)) {
-            throw InvalidParameterException.of(ErrorCode.PATH_IN_OUT_SAME);
+            throw InvalidParameterException.of(CommonErrorCode.PATH_IN_OUT_SAME);
         }
     }
