+++ src/main/java/nextstep/subway/station/ui/StationController.java
@@ -1,11 +1,8 @@
 package nextstep.subway.station.ui;
 
-import nextstep.subway.common.exception.ErrorCode;
-import nextstep.subway.common.exception.ErrorResponse;
 import nextstep.subway.station.application.StationService;
 import nextstep.subway.station.dto.StationRequest;
 import nextstep.subway.station.dto.StationResponse;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
